Security update - Aljoscha
The password to connect to MongoDB Atlas was initially included in the index.js to facilitate set up for other collaborators/reviewers at TechLabs. Removed the password and also changed it on MongoDB Atlas to prevent any future security issues, should other team members ever wish to continue the project.